requirements.txt
numpy==1.23.5
streamlit==1.22.0
Pillow==9.5.0
keras
tensorflow
matplotlib